+++ tests/integration/negotiate/test_401_response_missing_header.py
@@ -21,11 +21,12 @@
 # SOFTWARE.
 
 from multiprocessing import Process
+import sys
 from time import sleep
 
-from fastapi import Depends, FastAPI
-from fastapi.security import HTTPBasic, HTTPBasicCredentials
+from fastapi import FastAPI
 import pytest
+from starlette.requests import Request
 import uvicorn
 
 from ansys.openapi.common import (
@@ -34,49 +35,58 @@ from ansys.openapi.common import (
     AuthenticationScheme,
     SessionConfiguration,
 )
-
-from .integration.common import (
+from tests.integration.common import (
     TEST_MODEL_ID,
-    TEST_PASS,
     TEST_PORT,
-    TEST_URL,
-    TEST_USER,
     ExampleModelPyd,
     return_model,
-    validate_user_basic,
+    validate_user_principal,
 )
 
+pytestmark = pytest.mark.kerberos
+
+TEST_URL = f"http://test-server:{TEST_PORT}"
+TEST_PRINCIPAL = "httpuser@EXAMPLE.COM"
+
 custom_test_app = FastAPI()
-security = HTTPBasic()
+
+
+@custom_test_app.middleware("http")
+async def strip_www_authenticate_header(request: Request, call_next):
+    response = await call_next(request)
+    if response.status_code == 401:
+        del response.headers["www-authenticate"]
+    return response
 
 
 @custom_test_app.patch("/models/{model_id}")
-async def patch_model(
-    model_id: str,
-    example_model: ExampleModelPyd,
-    credentials: HTTPBasicCredentials = Depends(security),
-):
-    validate_user_basic(credentials)
+async def patch_model(model_id: str, example_model: ExampleModelPyd, request: Request):
+    validate_user_principal(request, TEST_PRINCIPAL)
     return return_model(model_id, example_model)
 
 
 @custom_test_app.get("/test_api")
-async def get_test_api(credentials: HTTPBasicCredentials = Depends(security)):
-    validate_user_basic(credentials)
+async def get_test_api(request: Request):
+    validate_user_principal(request, TEST_PRINCIPAL)
     return {"msg": "OK"}
 
 
 @custom_test_app.get("/")
-async def get_none(credentials: HTTPBasicCredentials = Depends(security)):
-    validate_user_basic(credentials)
+async def get_none(request: Request):
+    validate_user_principal(request, TEST_PRINCIPAL)
     return None
 
 
 def run_server():
-    uvicorn.run(custom_test_app, port=TEST_PORT)
+    # Function is only executed if testing in Linux
+    from asgi_gssapi import SPNEGOAuthMiddleware
+
+    authenticated_app = SPNEGOAuthMiddleware(custom_test_app, hostname="test-server")
+    uvicorn.run(authenticated_app, port=TEST_PORT)
 
 
-class TestBasic:
+@pytest.mark.skipif(sys.platform == "win32", reason="No portable KDC is available at present")
+class TestNegotiate:
     @pytest.fixture(autouse=True)
     def server(self):
         proc = Process(target=run_server, args=(), daemon=True)
@@ -86,36 +96,23 @@ class TestBasic:
         while proc.is_alive():
             sleep(1)
 
-    @pytest.mark.parametrize("auth_mode", [AuthenticationScheme.AUTO, AuthenticationScheme.BASIC])
-    def test_can_connect(self, auth_mode):
+    def test_can_connect(self):
         client_factory = ApiClientFactory(TEST_URL, SessionConfiguration())
-        _ = client_factory.with_credentials(
-            TEST_USER, TEST_PASS, authentication_scheme=auth_mode
+        _ = client_factory.with_autologon(
+            authentication_scheme=AuthenticationScheme.KERBEROS
         ).connect()
 
-    @pytest.mark.parametrize("auth_mode", [AuthenticationScheme.AUTO, AuthenticationScheme.BASIC])
-    def test_invalid_user_return_401(self, auth_mode):
+    def test_get_health_returns_200_ok(self):
         client_factory = ApiClientFactory(TEST_URL, SessionConfiguration())
-        with pytest.raises(ApiConnectionException) as exception_info:
-            _ = client_factory.with_credentials(
-                "eve", "password", authentication_scheme=auth_mode
-            ).connect()
-        assert exception_info.value.response.status_code == 401
-        assert "Unauthorized" in exception_info.value.response.reason
-
-    @pytest.mark.parametrize("auth_mode", [AuthenticationScheme.AUTO, AuthenticationScheme.BASIC])
-    def test_get_health_returns_200_ok(self, auth_mode):
-        client_factory = ApiClientFactory(TEST_URL, SessionConfiguration())
-        client = client_factory.with_credentials(
-            TEST_USER, TEST_PASS, authentication_scheme=auth_mode
+        client = client_factory.with_autologon(
+            authentication_scheme=AuthenticationScheme.KERBEROS
         ).connect()
 
         resp = client.request("GET", TEST_URL + "/test_api")
         assert resp.status_code == 200
         assert "OK" in resp.text
 
-    @pytest.mark.parametrize("auth_mode", [AuthenticationScheme.AUTO, AuthenticationScheme.BASIC])
-    def test_patch_model(self, auth_mode):
+    def test_patch_model(self):
         from . import models
 
         deserialized_response = models.ExampleModel(
@@ -126,7 +123,7 @@ class TestBasic:
         )
 
         resource_path = "/models/{ID}"
-        http_method = "PATCH"
+        method = "PATCH"
         path_params = {"ID": TEST_MODEL_ID}
 
         response_type = "ExampleModel"
@@ -134,17 +131,46 @@ class TestBasic:
         upload_data = {"ListOfStrings": ["red", "yellow", "green"]}
 
         client_factory = ApiClientFactory(TEST_URL, SessionConfiguration())
-        client = client_factory.with_credentials(
-            TEST_USER, TEST_PASS, authentication_scheme=auth_mode
+        client = client_factory.with_autologon(
+            authentication_scheme=AuthenticationScheme.KERBEROS
         ).connect()
         client.setup_client(models)
 
         response = client.call_api(
             resource_path,
-            http_method,
+            method,
             path_params=path_params,
             body=upload_data,
             response_type=response_type,
             _return_http_data_only=True,
         )
         assert response == deserialized_response
+
+
+@pytest.mark.skipif(sys.platform == "win32", reason="No portable KDC is available at present")
+class TestNegotiateFailures:
+    @pytest.fixture(autouse=True)
+    def server(self):
+        # Remove all the routes (a bit drastic)
+        custom_test_app.router.routes = []
+
+        @custom_test_app.get("/")
+        async def get_forbidden(request: Request):
+            validate_user_principal(request, "otheruser@EXAMPLE.COM")
+            return None
+
+        proc = Process(target=run_server, args=(), daemon=True)
+        proc.start()
+        yield
+        proc.terminate()
+        while proc.is_alive():
+            sleep(1)
+
+    def test_bad_principal_returns_403(self):
+        client_factory = ApiClientFactory(TEST_URL, SessionConfiguration())
+        with pytest.raises(ApiConnectionException) as excinfo:
+            _ = client_factory.with_autologon(
+                authentication_scheme=AuthenticationScheme.KERBEROS
+            ).connect()
+        assert excinfo.value.response.status_code == 403
+        assert excinfo.value.response.reason == "Forbidden"
